Code cleanup based on IDE (intellij) warnings.
Remove unused/commented-out code.

Use emptySetOf( Type ) instead of asSet().

Limit references to the deprecated GraphDatabaseSettings.

Remove usages of deprecated junit.framework package.

+++ community/kernel/src/test/java/org/neo4j/kernel/impl/traversal/TestBranchState.java
@@ -20,6 +20,7 @@
 package org.neo4j.kernel.impl.traversal;
 
 import org.junit.Test;
+
 import org.neo4j.graphdb.Direction;
 import org.neo4j.graphdb.Path;
 import org.neo4j.graphdb.PathExpander;
@@ -29,7 +30,7 @@ import org.neo4j.graphdb.traversal.Evaluation;
 import org.neo4j.graphdb.traversal.InitialBranchState;
 import org.neo4j.graphdb.traversal.PathEvaluator;
 
-import static junit.framework.Assert.assertEquals;
+import static org.junit.Assert.assertEquals;
 import static org.neo4j.graphdb.Direction.OUTGOING;
 import static org.neo4j.graphdb.traversal.Evaluation.ofIncludes;
 import static org.neo4j.helpers.collection.IteratorUtil.count;
@@ -94,7 +95,7 @@ public class TestBranchState extends AbstractTestBase
         {
             assertEquals( path.length()/2, state.getState().intValue() );
             if ( path.length() % 2 == 1 )
-                state.setState( ((Integer) state.getState())+1 );
+                state.setState( state.getState() + 1 );
             return path.endNode().getRelationships( Direction.OUTGOING );
         }
